Added  Events history API and fixed Issue

+++ src/app/admin/customer/add-file/add-file.component.ts
@@ -1,3 +1,4 @@
+import { DatePipe } from '@angular/common';
 import { Component, ElementRef, Inject, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
@@ -17,6 +18,10 @@ export class AddFileComponent implements OnInit, OnDestroy {
   userId: any;
   customerFileList: any = [];
   showDelete = false;
+  showDownload = false;
+  showNew = false;
+  showView = false;
+  errorMessage: any;
   index = 0;
   isList = false;
   fileObject: any;
@@ -25,6 +30,7 @@ export class AddFileComponent implements OnInit, OnDestroy {
     private readonly formBuilder: FormBuilder,
     private readonly customerService: CustomerService,
     private readonly el: ElementRef,
+    private readonly datePipe: DatePipe,
     public dialogRef: MatDialogRef<AddFileComponent>,
     @Inject(MAT_DIALOG_DATA) public data: any) {
     this.userId = JSON.parse(localStorage.getItem('users')).userId;
@@ -41,6 +47,9 @@ export class AddFileComponent implements OnInit, OnDestroy {
       .subscribe((customerFile: any) => {
         if (customerFile.customerManagement.customerFileList.list.length > 0) {
           this.showDelete = true;
+          this.showDownload = true;
+          this.showView = true;
+          this.showNew = true;
           this.customerFileList = customerFile.customerManagement.customerFileList.list;
           this.setForm(customerFile.customerManagement.customerFileList.list[0]);
         } else {
@@ -52,10 +61,16 @@ export class AddFileComponent implements OnInit, OnDestroy {
   }
 
   onNoClick() {
-    this.dialogRef.close(false);
+    this.dialogRef.close(true);
   }
 
   handle(files: any) {
+    this.showDelete = false;
+    this.showDownload = false;
+    this.showView = false;
+    this.showNew = false;
+    this.customerFileList = [];
+    this.setForm(undefined);
     if (files.length > 0) {
       this.fileObject = files[0];
       this.customerFileForm.controls['fileName'].setValue(files[0].name);
@@ -70,7 +85,7 @@ export class AddFileComponent implements OnInit, OnDestroy {
     }
     this.customerFileForm = this.formBuilder.group({
       id: [event !== undefined ? event.id : ''],
-      timestamp: [event !== undefined ? new Date(event.timestamp) : new Date()],
+      timestamp: [event && event.timestamp ? this.datePipe.transform(new Date(event.timestamp), 'MM/dd/yyyy') : this.datePipe.transform(new Date(), 'MM/dd/yyyy')],
       fileName: [event !== undefined ? event.fileName : '', Validators.required],
       originalFileName: [event !== undefined ? event.originalFileName : ''],
       size: [event !== undefined ? event.size : ''],
@@ -101,27 +116,26 @@ export class AddFileComponent implements OnInit, OnDestroy {
     }
   }
   newEvent() {
-    this.showDelete = false;
-    this.customerFileList = [];
-    this.setForm(undefined);
     const element: HTMLElement = document.querySelectorAll('#fileInput')[0] as HTMLElement;
     element.click();
   }
 
   delete() {
-    this.subscriptions.add(this.customerService.deleteCustomerFileById(this.data.customerId, this.customerFileForm.controls.fileName.value).pipe(skipWhile((item: any) => !item))
-      .subscribe((response: any) => {
-        this.dialogRef.close(true);
-      }));
+    if (confirm('Are you sure you want to delete?')) {
+      this.subscriptions.add(this.customerService.deleteCustomerFileById(this.data.customerId, this.customerFileForm.controls.fileName.value).pipe(skipWhile((item: any) => !item))
+        .subscribe((response: any) => {
+          this.dialogRef.close(true);
+        }));
+    }
   }
 
   download() {
-    const url = window.location.origin + '/' + this.customerFileForm.controls.originalFileName.value + '?preview=false&fileName=Test.txt';
+    const url = window.location.origin + '/hea-web/' + this.customerFileForm.controls.originalFileName.value + '?preview=false&fileName=' + this.customerFileForm.controls.fileName.value;
     window.open(url, '_blank');
   }
 
   view() {
-    const url = window.location.origin + '/' + this.customerFileForm.controls.originalFileName.value + '?preview=true&fileName=Test.txt';
+    const url = window.location.origin + '/hea-web/' + this.customerFileForm.controls.originalFileName.value + '?preview=true&fileName=' + this.customerFileForm.controls.fileName.value;
     window.open(url, '_blank');
   }
   save() {
@@ -130,7 +144,12 @@ export class AddFileComponent implements OnInit, OnDestroy {
         this.subscriptions.add(this.customerService.updateCustomerFile(this.data.customerId, this.data.row.id,
           { 'fileName': this.customerFileForm.controls.fileName.value, 'description': this.customerFileForm.controls.description.value }).pipe(skipWhile((item: any) => !item))
           .subscribe((response: any) => {
-            this.dialogRef.close(true);
+            // this.dialogRef.close(true);
+            if (response.customerManagement.customerFile.errorMessage) {
+              this.errorMessage = response.customerManagement.customerFile.errorMessage;
+            } else {
+              this.loadCustomerFile(this.data.customerId);
+            }
           }));
       } else {
         let description = '';
@@ -139,7 +158,11 @@ export class AddFileComponent implements OnInit, OnDestroy {
         }
         this.subscriptions.add(this.customerService.saveCustomerFile(this.data.customerId, this.fileObject, description).pipe(skipWhile((item: any) => !item))
           .subscribe((response: any) => {
-            this.dialogRef.close(true);
+            if (response.customerManagement.customerFile.errorMessage) {
+              this.errorMessage = response.customerManagement.customerFile.errorMessage;
+            } else {
+              this.loadCustomerFile(this.data.customerId);
+            }
           }));
       }
     } else {
